Fix some races in the executor
Both are pretty unlikely but worth fixing:

1/ The executors cache needs to be thread safe. Since we're only
testing against a single database, once the executor is instantiated
and cached successfully, the race can't manifest itself anymore.

2/ Fix a static destruction order fiasco: llvm::ExecutionEngine depends
on static state in LLVM -- a mutex. By the time execution tests nuke the
executors cache that mutex doesn't exist anymore. This is not an issue
for the server, just for the test. But it cleans up the output of
-fsanitize=thread.