Don't hold the lock until return in sql_execute
Turns out the copy command goes through it (as opposed to load_table
I've previously fixed) and it's not a good idea anyway since the
server will need to reject connections.

In order to be thread safe without holding the lock we need to copy
the SessionInfo object, but this shouldn't be a problem.