Fix overly conservative bail in Executor::skipFragment
A recent change in how types are compared (2dbdda3) means that nullable and
not nullable types which are identical otherwise are not considered equal
anymore. Executor::skipFragment bailed when left and right types were not
equal, which is overly conservative but was hidden before 2dbdda3 exposed
this problem. Types can only differ in nullability for simple qualifiers
used for skipping fragments and it's ok if they're not the same in that
regard. Remove that check, since it's both redundant and too conservative.