web: provide servers.json in v2 subdir
Requests for `/v2/servers.json` will now return the same file as
`/servers.json` if `<frontend>/v2/servers.json` does not exist on disk.

This is required because we currently provide the Immerse V2 beta in a
subdirectory of the main V1 frontend. This worked previously because V2
loaded the absolute `/servers.json` instead of the relative
`servers.json`. The old behavior is incorrect if different versions of
`servers.json` are required by different frontends on the same hostname,
as is the case for www.mapd.com/demos/*.

cc: map-d/mapd2-frontend#1541