Add has_cardinality_estimation parameter to Executor::executeWorkUnit
Whether we use the baseline hash or not is decided at a fairly low level.
But now we'll also need that information to decide whether we should run
a cardinality estimation query first. Add a parameter which tells the
lower layers a cardinality estimation is available on request, initially
false. The lower level will throw an exception to request it.