Make leaves execute a step even if input is replicated
If the query involves working on a mix of replicated and sharded inputs, leaves
will go out of sync if they didn't execute the replicated steps at all. Execute
the step anyway and let the aggregator discard all but one results. This is
largely harmless since leaves work in parallel anyway, plus it's a rare case.