Short-circuit query filters on unlikely quals, issue #574
Added support for top level filter short-circuiting, akin to how codegen
short-circuits logical exprs but dealing with multiple lists of disjoint
quals: simple quals, quals, inner join quals, deferred quals.

Any quals that are unlikely to be true (< 10%), particularly quals that
use UNLIKELY hints directly, will be prioritized and checked first.
All other quals will be deferred and would be checked only if the unlikely
quals unexpectedly turn out to be true. The weight of the short-circuited
quals is not taken into account, we assume users know what they're doing.

Other things to note:
No change in behavior if none of the quals are determined to be unlikely or
if there is nothing to short-circuit. Unlikely quals that contain unsafe
division are not brought forward.