Replace --disable-rendering with --rendering=false
Previously the value provided on the command line/in the config file was
passed directly to the internal `enable_rendering` var, meaning that in
order to disable rendering you would do `--disable-rendering=false`
instead of `=true`. Now `--rendering=false` will disable rendering.

Also removed unnecessary `--disable-fork` check and moved deprecated
option check to be the first output in the logs.